requirements.txt
numpy==1.24.4
Pillow==10.3.0
streamlit==1.36.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
